Reorganized streamfow assimilation functions
