+++ imports/streamflow_assimilation.py
@@ -7,11 +7,13 @@ import os
 from pytz import utc
 import requests
 
+from RAPIDpy.rapid import RAPID
 from RAPIDpy.dataset import RAPIDDataset
 from RAPIDpy.helper_functions import csv_to_list
 
 
 from helper_functions import get_ensemble_number_from_forecast
+
 #-----------------------------------------------------------------------------------------------------
 # StreamSegment Class
 #-----------------------------------------------------------------------------------------------------
@@ -329,22 +331,81 @@ class StreamNetworkInitializer(object):
                     init_flow_file.write("{}\n".format(stream_segment.init_flow))
         
         
-        
-        
-if __name__=="__main__":
-        connect_file = '/home/alan/work/rapid-io/input/erdc_texas_gulf_region-huc_2_12/rapid_connect.csv'
-        gage_flow_info = '/home/alan/work/rapid-io/input/erdc_texas_gulf_region-huc_2_12/usgs_gages.csv'
-        sni = StreamNetworkInitializer(connectivity_file=connect_file, gage_ids_natur_flow_file=gage_flow_info)
-        path_to_predictions = '/home/alan/tethysdev/tethysapp-erfp_tool/rapid_files/ecmwf_prediction/erdc_texas_gulf_region/huc_2_12/20150825.1200'
-        prediction_files = glob(os.path.join(path_to_predictions, "*.nc"))
+#-----------------------------------------------------------------------------------------------------
+# Streamflow Init Functions
+#-----------------------------------------------------------------------------------------------------
+def compute_initial_rapid_flows(prediction_files, input_directory, forecast_date_timestep):
+    """
+    Gets mean of all 52 ensembles 12-hrs in future and prints to csv as initial flow
+    Qinit_file (BS_opt_Qinit)
+    The assumptions are that Qinit_file is ordered the same way as rapid_connect_file
+    if subset of list, add zero where there is no flow
+    """
+    #remove old init files for this basin
+    past_init_flow_files = glob(os.path.join(input_directory, 'Qinit_*.csv'))
+    for past_init_flow_file in past_init_flow_files:
+        try:
+            os.remove(past_init_flow_file)
+        except:
+            pass
+    current_forecast_date = datetime.datetime.strptime(forecast_date_timestep[:11],"%Y%m%d.%H")
+    current_forecast_date_string = current_forecast_date.strftime("%Y%m%dt%H")
+    init_file_location = os.path.join(input_directory,'Qinit_%s.csv' % current_forecast_date_string)
+    #check to see if exists and only perform operation once
+    if prediction_files:
+        sni = StreamNetworkInitializer(connectivity_file=os.path.join(input_directory,'rapid_connect.csv'))
         sni.compute_init_flows_from_past_forecast(prediction_files)
-        raw_initialization_file =  '/home/alan/work/rapid-io/input/erdc_texas_gulf_region-huc_2_12/Qinit_20150825t12_orig.csv'
-        sni.write_init_flow_file(raw_initialization_file)
-        #sni.read_init_flows_from_past_forecast(raw_initialization_file)
-        sni.add_usgs_flows(datetime.datetime(2015,8,26,0, tzinfo=utc))
+        sni.write_init_flow_file(init_file_location)        
+    else:
+        print "No current forecasts found. Skipping ..."
+
+def compute_seasonal_initial_rapid_flows(historical_qout_file, input_directory, forecast_date_timestep):
+    """
+    Gets the seasonal average from historical file to initialize from
+    """
+    current_forecast_date = datetime.datetime.strptime(forecast_date_timestep[:11],"%Y%m%d.%H")
+    #move the date back a forecast (12 hrs) to be used in this forecast
+    forecast_date_string = (current_forecast_date-datetime.timedelta(seconds=12*3600)).strftime("%Y%m%dt%H")
+    init_file_location = os.path.join(input_directory,'Qinit_%s.csv' % forecast_date_string)
+    if not os.path.exists(init_file_location):
+        #check to see if exists and only perform operation once
+        if historical_qout_file and os.path.exists(historical_qout_file):
+            rapid_manager = RAPID(Qout_file=historical_qout_file,
+                                  rapid_connect_file=os.path.join(input_directory,'rapid_connect.csv'))
+            rapid_manager.generate_seasonal_intitialization(init_file_location)
+        else:
+            print "No seasonal streamflow file found. Skipping ..."
+
+def compute_seasonal_initial_rapid_flows_multicore_worker(args):
+    """
+    Worker function using mutliprocessing for compute_seasonal_initial_rapid_flows
+    """
+    compute_seasonal_initial_rapid_flows(args[0], args[1], args[2])
+    
+def update_inital_flows_usgs(input_directory, forecast_date_timestep):
+    """
+    Update initial flows with USGS data
+    """
+    gage_flow_info = os.path.join(input_directory, 'usgs_gages.csv')
+    current_forecast_date = datetime.datetime.strptime(forecast_date_timestep[:11],"%Y%m%d.%H").replace(tzinfo=utc)
+    past_date = (datetime.datetime.strptime(forecast_date_timestep[:11],"%Y%m%d.%H") - \
+                 datetime.timedelta(hours=12)).replace(tzinfo=utc).strftime("%Y%m%dt%H")
+
+    qinit_file = os.path.join(input_directory, 'Qinit_%s.csv' % past_date)
+
+    if os.path.exists(gage_flow_info) and os.path.exists(qinit_file):
+        print "Updating initial flows with USGS data for:", \
+              input_directory, forecast_date_timestep , "..."
+              
+        sni = StreamNetworkInitializer(connectivity_file=os.path.join(input_directory,'rapid_connect.csv'),
+                                       gage_ids_natur_flow_file=gage_flow_info)
+        sni.read_init_flows_from_past_forecast(qinit_file)
+        sni.add_usgs_flows(current_forecast_date)
         sni.modify_init_flows_from_gage_flows()
-        #usgs_initialization_file =  '/Users/rdchlads/Documents/nfie_texas_gulf_initialization_test/usgs_init.csv'
-        #sni.write_init_flow_file(usgs_initialization_file)        
-        usgs_mod_initialization_file =  '/home/alan/work/rapid-io/input/erdc_texas_gulf_region-huc_2_12/Qinit_20150825t12_usgs.csv'
-        sni.write_init_flow_file(usgs_mod_initialization_file)        
+        try:
+            os.remove(qinit_file)
+        except OSError:
+            pass
+        
+        sni.write_init_flow_file(qinit_file)
